rework main

+++ src/main/java/Main.java
@@ -1,11 +1,11 @@
 public class Main {
     public static void main(String[] args) {
 
-        float price = 1020f;//стоимостиь билета
-        float amount = 20f; // количество
+        int price = 1020;//стоимостиь билета
+        int amount = 20; // количество
         int m = 1; //мили
 
-        float bonus = (price / amount) * m;
+        int bonus = (price / amount) * m;
 
         System.out.println(bonus);
     }
